requirements.txt
flask
pathlib
gunicorn
mysql.connector
python-dotenv
